@@ -11,7 +11,8 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Versioning;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Swashbuckle.AspNetCore.Swagger;
+using Microsoft.Extensions.Hosting;
+using Microsoft.OpenApi.Models;
 
 namespace ESFA.DC.LARS.API
 {
@@ -29,7 +30,7 @@ namespace ESFA.DC.LARS.API
         // This method gets called by the runtime. Use this method to add services to the container.
         public IServiceProvider ConfigureServices(IServiceCollection services)
         {
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_3_0);
 
             services.AddApiVersioning(options =>
             {
@@ -59,20 +60,21 @@ namespace ESFA.DC.LARS.API
             // Register the Swagger generator, defining 1 or more Swagger documents
             services.AddSwaggerGen(c =>
             {
-                c.SwaggerDoc("v1", new Info { Title = "Learning Aim Reference Service API", Version = "v1" });
+                c.SwaggerDoc("v1", new OpenApiInfo { Title = "Learning Aim Reference Service API", Version = "v1" });
             });
 
             return ConfigureAutofac(services);
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
             }
 
+            app.UseRouting();
             app.UseSwagger();
 
             app.UseSwaggerUI(c =>
@@ -80,7 +82,13 @@ namespace ESFA.DC.LARS.API
                 c.SwaggerEndpoint("/swagger/v1/swagger.json", "Learning Aim Reference Service API");
             });
 
-            app.UseMvc();
+            app.UseEndpoints(routes =>
+            {
+                routes.MapControllerRoute(
+                             name: "areaRoute",
+                             pattern: "{area:exists}/{controller}/{action}/{id?}",
+                             defaults: new { action = "Index" });
+            });
         }
 
         private IServiceProvider ConfigureAutofac(IServiceCollection services)
