@@ -1,4 +1,7 @@
-﻿<footer class="govuk-footer " role="contentinfo">
+﻿
+@inject IAppVersionService AppVer
+
+<footer class="govuk-footer " role="contentinfo">
     <div class="govuk-width-container ">
 
         <div class="govuk-footer__meta">
@@ -16,6 +19,10 @@
                 <a class="govuk-footer__link govuk-footer__copyright-logo" href="https://www.nationalarchives.gov.uk/information-management/re-using-public-sector-information/uk-government-licensing-framework/crown-copyright/">© Crown copyright</a>
             </div>
         </div>
-    </div>
+        <div class="govuk-footer__meta">
+            <div class="govuk-footer__meta-item govuk-footer__meta-item--grow">
+                (Version : @AppVer.Version - Runtime : @AppVer.TargetFrameworkVersion)
+            </div>
+        </div>
 </footer>
 <div id="global-app-error" class="app-error hidden"></div>
